+++ src/main/java/tn/esprit/spring/AhmedGuedri/Repositories/FeedObjectRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.AhmedGuedri.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.AhmedGuedri.entities.FeedObject;
 
-public interface FeedObjectRepository extends JpaRepository<FeedObject, Integer> {
+public interface FeedObjectRepository extends JpaRepository<FeedObject, Long> {
 }
