+++ src/main/java/tn/esprit/spring/AhmedGuedri/Repositories/FeedbacksRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.AhmedGuedri.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.AhmedGuedri.entities.Feedbacks;
 
-public interface FeedbacksRepository extends JpaRepository<Feedbacks, Integer> {
+public interface FeedbacksRepository extends JpaRepository<Feedbacks, Long> {
 }
