+++ src/main/java/tn/esprit/spring/AhmedGuedri/Repositories/InquiryRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.AhmedGuedri.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.AhmedGuedri.entities.Inquiry;
 
-public interface InquiryRepository extends JpaRepository<Inquiry, Integer> {
+public interface InquiryRepository extends JpaRepository<Inquiry, Long> {
 }
