+++ src/main/java/tn/esprit/spring/AhmedGuedri/Repositories/InvoicesRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.AhmedGuedri.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.AhmedGuedri.entities.Invoices;
 
-public interface InvoicesRepository extends JpaRepository<Invoices, Integer> {
+public interface InvoicesRepository extends JpaRepository<Invoices, Long> {
 }
